Allow handlers to use the strong environment
Now allows handlers to use the AppEnv (strong) if required, instead of the weak variant without a big chance of having bugs through  casting as only this parameter gets cast.
Hono's middleware will always pass the correct env, but it does not know it. This is an issue with the hono event emitter middleware libary. But fixed for the forseeable future. :)